pyproject.toml
[tool.poetry]
name = "candle_prediction_bot"
version = "0.1.0"
description = ""
authors = ["Vladislav Ozhgibesov <vozhgibesov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
numpy = "*"
scikit-learn = "*"
tensorflow = "*"
keras = "*"
python-telegram-bot = "*"
requests = "*"
websocket-client = "*"
tensorflow-io-gcs-filesystem = "*"
tradingview_ta = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-bot = "tg_bot:app.run_polling"    